requirements.txt
# Core dependencies
streamlit==1.22.0
tensorflow==2.10.1
numpy==1.24.4
pandas==1.5.3
pillow==10.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86

# Visualization
plotly==6.3.1
matplotlib==3.9.0
seaborn==0.13.2

# Image processing
scikit-image==0.20.0
imutils==0.5.4

# Utilities
python-dateutil==2.9.0.post0